Refactor to reduce duplicated code
